Adjusted all the constraints.Continue TO READ THIS
I have adjusted the constraints from absolute values into relative ones. HOWEVER, I have adjusted most of the display. NOT PRECISELY ACCORDING TO THE FIGMA. I have actually made a lot of adjustments to it. Because I simply could not figure out how to achieve that. The question is mainly that, when I transform the absolute values  of the fonts from Figma into relative ones, the result seems to be improper. As can be read from my code, I actually manually set all the proportion...
Moreover, THERE ARE PROBLEMS REGARDING THE TIME. how can we set the time?